# api/app/Http/Controllers/Auth/AuthController.php
<?php

namespace App\Http\Controllers\Auth;

use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Hash;
use App\Http\Controllers\Controller;
use App\Http\Requests\AuthLoginRequest;


class AuthController extends Controller
{

    /**
    * @OA\Post(
    * path="/api/register",
    * summary="register",
    * description="register by email, password",
    * operationId="register",
    * tags={"auth"},
    * @OA\RequestBody(
    *    required=true,
    *    description="Pass user credentials",
    *    @OA\JsonContent(
    *       required={"email","password", "persistent"},
    *       @OA\Property(property="email", type="string", format="email", example="user1@mail.com"),
    *       @OA\Property(property="password", type="string", format="password", example="PassWord12345"),
    *       @OA\Property(property="persistent", type="boolean", example=true),
    *    ),
    * ),
    * @OA\Response(
    *    response=422,
    *    description="Wrong credentials response",
    *    @OA\JsonContent(
    *       @OA\Property(property="message", type="string", example="Sorry, wrong email address or password. Please try again")
    *    )
    * )
    * )
    */
    public function createUser(Request $request) 
    {
        try {
            User::create([
                'name' => $request->name,
                'email' => $request->email,
                'password' => Hash::make($request->password)
            ]);

            return response()->json([
                'status' =>  true,
                'msg' => 'usuário cadastrado com sucesso',
            ], 200);

        } catch(\Exception $e) { 
            return response()->json([
                'status' => false,
                'msg' => $e->getMessage()
            ], 500);
        }
    }

    /**
    * @OA\Post(
    * path="/api/login",
    * summary="Sign in",
    * description="Login by email, password",
    * operationId="authLogin",
    * tags={"auth"},
    * @OA\RequestBody(
    *    required=true,
    *    description="Pass user credentials",
    *    @OA\JsonContent(
    *       required={"email","password", "persistent"},
    *       @OA\Property(property="email", type="string", format="email", example="user1@mail.com"),
    *       @OA\Property(property="password", type="string", format="password", example="PassWord12345"),
    *       @OA\Property(property="persistent", type="boolean", example=true),
    *    ),
    * ),
    * @OA\Response(
    *    response=422,
    *    description="Wrong credentials response",
    *    @OA\JsonContent(
    *       @OA\Property(property="message", type="string", example="Sorry, wrong email address or password. Please try again")
    *    )
    * )
    * )
    */
    public function login(AuthLoginRequest $request) 
    {

        $credentials = $request->only('email', 'password');

        if (Auth::attempt($credentials)) {
            $request->session()->regenerate();

            $user = Auth::user();

            $token = $user->createToken('auth_token')->plainTextToken;

            return response()->json(['user' =>  $user, 'token' => $token], 200);
        }
    }



    /**
     * @OA\Post(
     * path="/api/logout",
     * summary="Sign out",
     * description="Logout user",
     * operationId="authLogout",
     * tags={"auth"},
     * security={ {"bearer": {} }},
     * @OA\Response(
     *    response=200,
     *    description="Successful logout",
     *    @OA\JsonContent(
     *       @OA\Property(property="message", type="string", example="Logged out successfully")
     *    )
     * ),
     * @OA\Response(
     *    response=401,
     *    description="Unauthorized",
     *    @OA\JsonContent(
     *       @OA\Property(property="message", type="string", example="Unauthorized")
     *    )
     * )
     * )
     */
    public function logout(Request $request) 
    {
        Auth::guard('sanctum')->user()->tokens()->delete();

        $request->session()->invalidate();

        $request->session()->regenerateToken();
    }
}